[531] more refactoring after merge with Anzette work
