+++ src/modules/land-works/components/lands-explore-list/index.tsx
@@ -3,7 +3,6 @@ import { useHistory, useLocation } from 'react-router-dom';
 import useMediaQuery from '@mui/material/useMediaQuery';
 import useDebounce from '@rooks/use-debounce';
 
-import { AtlasTile } from 'components/custom/atlas';
 import { Grid } from 'design-system';
 import { GridIcon, MapIcon } from 'design-system/icons';
 import { LocationState } from 'modules/interface';
@@ -16,6 +15,7 @@ import { useLandsMapTile } from 'modules/land-works/providers/lands-map-tile';
 import { useLandsMapTiles } from 'modules/land-works/providers/lands-map-tiles';
 import { useLandsSearchQuery } from 'modules/land-works/providers/lands-search-query';
 
+import { AtlasTile } from '../atlas';
 import {
   LandsSearchBarWrapperStyled,
   StyledButton,
